Ex2: Added changes to the avatar bades on the arrow keys
The avatar is invisible until the player pushes the arrow keys. When pressing the left arrow, the avatar is blue with a green outline. When pressing the right arrow, the avatar is green with a blue outline. When pressing the up arrow, the avatar is yellow with a red outline. When pressing the down arrow, the avatar is red with a yellow outline.