P3: Fixed Enemy and Modified Prey
The radius of the enemy did not match the position of the image. By having imageMode(CENTER), the image and radius finally watch and no longer causes confusion.

The Prey's stroke shrinks in size to make it more dynamic. In addition, all of the prey's have a different time interval so that they don't all move at the same time.