P3: Added reset function
When the game ends, the player will be given the option to continue the game by pressing the spacebar. The predator, prey and enemy will reset position, speed and heath. The only thing that will not reset are the points bacause i want the player to continue collecting points even though they will never win regardless of the amount.